Move the useLibrary declaration from a root extension to the embeddedTools extension
By doing this, they are now type safe. This makes them easier to use from Java, and to extend on if necessary.
It also avoids poluting the root extension space. Its a slight breaking change, but in gradlerio we can actually just add duplicates on the wpi extension.

That would make in C++ the use library change to wpi.useLibrary(it, ...) which is much cleaner, and intellisense would work