Fix single missing toolchain print
The logic was backwards before, and after thinking and testing, the flag in ET should just skip all binary prints
In the TC plugin, we had an additional print done by the GccToolChain. That should be the one set to single print,
and this one should be completely disableable, as the other one makes much more sense to an end user.